refactor: moves api helper functions to api/utils
